feat(entrypoint): added entrypoint.extract functionality to support extensions and additional configuration files for firefly

+++ docker/entrypoint.py
@@ -13,7 +13,8 @@ from typing import List
 
 SLIDES: str = "\n" + ("=" * 40) + "\n"
 
-def extract(source: Path, destination: Path, prefix:str = "/", kind: str = "war"):    
+
+def extract(source: Path, destination: Path, prefix: str = "/", kind: str = "war"):
     """Extract a file to a destination directory.
 
     Args:
@@ -21,54 +22,67 @@ def extract(source: Path, destination: Path, prefix:str = "/", kind: str = "war"
         destination (Path): Destination directory to extract to
         prefix (str, optional): Prefix extracted filepaths. Default is "/".
         kind (str, optional): File extension to extract. Default is "war".
+
+    Raises:
+        FileNotFoundError: If source directory does not exist.
     """
-    print(f"{SLIDES} Extracting {kind.upper()} files {SLIDES}")
     extractions: dict[str, Path] = {}
+    print(f"{SLIDES} Extracting Files {SLIDES}")
+    print(f"Source     : {source}")
+    print(f"Destination: {destination}")
+    print(f"Prefix     : {prefix}")
+    print(f"Kind       : {kind}")
+    print("\n")
+
+    if not source.exists() or not source.is_dir():
+        raise FileNotFoundError(f"Source directory does not exist: {source}")
+
     for item in source.iterdir():
         # Only extract specified file extensions
         if item.is_file() and item.suffix == f".{kind}":
             filename = item.stem
-
+            print(f"Extracting: {item}")
             # Configure prefix for extracted files
             prefix = prefix.replace("/", "#").strip("#")
             if prefix:
-                filename = f"{prefix}#{filename}"
+                landing = destination / f"{prefix}#{filename}"
+            else:
+                landing = destination / filename
 
-            # Create landing directory for extracted files
-            landing = destination / filename      
+            # Create landing directory if it does not exist
+            print(f"    Preparing -> {landing}")
             landing.mkdir(parents=True, exist_ok=True)
-            
             # Extract archive to landing directory
-            print(f"Extracting {item} to {landing}")
+            print(f"    Extracting {item} -> {landing}")
             with ZipFile(item, "r") as archive:
                 archive.extractall(landing)
-            print(f"Extracted {item} to {landing}")
             extractions[filename] = landing
-            
             # Modify log4j to send logs to stdout
             log4j = destination / "WEB-INF/classes/log4j2.properties"
             if log4j.exists():
                 # Create backup as log4j.properties.bak
-                print(f"Modifying {log4j}")
+                print(f"    Modifying {log4j}")
                 shutil.copy(log4j, log4j.with_suffix(".bak"))
-                with open(log4j, 'r+') as file:
+                with open(log4j, "r+", encoding="utf-8") as file:
                     content = file.read()
                     content = content.replace("##out--", "")
                     file.seek(0)
                     file.write(content)
                     file.truncate()
-    
+
     # Copy secondary files for extracted archives to landing directory
     for filename, landing in extractions.items():
         origin = source / f"{filename}"
-        for item in origin.iterdir():
-            if item.is_dir():
-                print(f"Copying directory {item} to {landing}")
-                shutil.copytree(item, landing / item.name, dirs_exist_ok=True)
-            else:
-                print(f"Copying file {item} to {landing}")
-                shutil.copy2(item, landing / item.name)
-    print(f"{SLIDES} Extractions complete {SLIDES}")
+        if origin.exists() and origin.is_dir():
+            print(f"Post Init: {origin}")
+            for item in origin.iterdir():
+                if item.is_dir():
+                    print(f"    Copying tree {item} -> {landing}")
+                    shutil.copytree(item, landing / item.name, dirs_exist_ok=True)
+                else:
+                    print(f"    Copying file {item} -> {landing}")
+                    shutil.copy2(item, landing / item.name)
+    print(f"{SLIDES} Extractions Done {SLIDES}")
 
 
 def add_multi_props_env_var():
@@ -118,16 +132,17 @@ def add_single_prop_env_vars():
     return props_opts
 
 
-def log_env_info(path_prefix, visualize_fits_search_path):
+def log_env_info(path_prefix: str, visualize_fits_search_path: str):
     """Log environment variables and configuration information."""
-    print(f"""
+    print(
+        f"""
     ========== Information: You can set environment variables using -e on the docker run line =====
 
     Environment Variables:
             Description                      Name                          Value
             -----------                      --------                      -----
             Admin username                   ADMIN_USER                    {os.getenv('ADMIN_USER')}
-            Admin password                   ADMIN_PASSWORD                {os.getenv('ADMIN_PASSWORD')}
+            Admin password          s         ADMIN_PASSWORD                {os.getenv('ADMIN_PASSWORD')}
             Additional data path             VISUALIZE_FITS_SEARCH_PATH    {visualize_fits_search_path}
             Clean internal (e.g., 720m, 5h)  CLEANUP_INTERVAL              {os.getenv('CLEANUP_INTERVAL', '')}
             Context path prefix              PATH_PREFIX                   {path_prefix}
@@ -158,28 +173,34 @@ def log_env_info(path_prefix, visualize_fits_search_path):
     Command line options:
             --help  : show help message, examples, stop
             --debug : start in debug mode
-    """)
+    """
+    )
 
 
-def show_help(name:str, webapps_ref: Path):
+def show_help(name: str, webapps_ref: Path):
     """Show help message and examples.
 
     Args:
         name (str): Application name
         webapps_ref (Path): Reference webapps directory
     """
-    with open("./start-examples.txt") as f:
+    with open("./start-examples.txt", encoding="utf-8") as f:
         print(f.read().replace("ipac/firefly", name))
 
     for item in webapps_ref.iterdir():
-        if item.is_file() and item.stem == "firefly" and item.suffix == ".war" :
-            with open("./customize-firefly.txt") as f:
+        if item.is_file() and item.stem == "firefly" and item.suffix == ".war":
+            with open("./customize-firefly.txt", encoding="utf-8") as f:
                 print(f.read())
     sys.exit(0)
 
+
 def dry_run(cmd: List[str], webapps: Path):
-    """Display a dry run of the command and environment setup."""
-    
+    """Display a dry run of the command and environment setup.
+
+    Args:
+        cmd (List[str]): Command to run
+        webapps (Path): Webapps directory
+    """
     print(f"\n{SLIDES} DRY RUN {SLIDES}")
     print(f"COMMAND      : {' '.join(cmd)}")
     print(f"CATALINA_OPTS: {os.getenv('CATALINA_OPTS')}\n")
@@ -203,29 +224,40 @@ def main():
     start_mode = os.getenv("START_MODE", "run")
     name = os.getenv("BUILD_TIME_NAME", "ipac/firefly")
     admin_user = os.getenv("ADMIN_USER", "admin")
-    admin_password = os.getenv("ADMIN_PASSWORD", base64.b64encode(str(random.randint(100000, 999999)).encode()).decode()[:8])
+    admin_password = os.getenv(
+        "ADMIN_PASSWORD",
+        base64.b64encode(str(random.randint(100000, 999999)).encode()).decode()[:8],
+    )
     use_admin_auth = os.getenv("USE_ADMIN_AUTH", "true").lower()
-    path_prefix = os.getenv("PATH_PREFIX") or os.getenv("baseURL") or ""      # use PATH_PREFIX; baseURL is for backward compatibility
+    path_prefix = (
+        os.getenv("PATH_PREFIX") or os.getenv("baseURL") or ""
+    )  # use PATH_PREFIX; baseURL is for backward compatibility
 
-    vis_path = "/external" if not visualize_fits_search_path \
+    vis_path = (
+        "/external"
+        if not visualize_fits_search_path
         else f"/external:{visualize_fits_search_path}"
+    )
 
-    catalina_opts = " ".join([
-        f"-XX:InitialRAMPercentage={os.getenv('INIT_RAM_PERCENT', '10')}",
-        f"-XX:MaxRAMPercentage={os.getenv('MAX_RAM_PERCENT', '80')}",
-        "-XX:+UnlockExperimentalVMOptions", "-XX:TrimNativeHeapInterval=30000",
-        f"-DADMIN_USER={admin_user}",
-        f"-DADMIN_PASSWORD={admin_password}",
-        f"-Dhost.name={os.getenv('HOSTNAME', '')}",
-        f"-Dserver.cores={os.getenv('JVM_CORES', '')}",
-        "-Djava.net.preferIPv4Stack=true",
-        "-Dwork.directory=/firefly/workarea",
-        "-Dshared.work.directory=/firefly/shared-workarea",
-        "-Dserver_config_dir=/firefly/config",
-        "-Dstats.log.dir=/firefly/logs/statistics",
-        "-Dalerts.dir=/firefly/alerts",
-        f"-Dvisualize.fits.search.path={vis_path}"
-    ])
+    catalina_opts = " ".join(
+        [
+            f"-XX:InitialRAMPercentage={os.getenv('INIT_RAM_PERCENT', '10')}",
+            f"-XX:MaxRAMPercentage={os.getenv('MAX_RAM_PERCENT', '80')}",
+            "-XX:+UnlockExperimentalVMOptions",
+            "-XX:TrimNativeHeapInterval=30000",
+            f"-DADMIN_USER={admin_user}",
+            f"-DADMIN_PASSWORD={admin_password}",
+            f"-Dhost.name={os.getenv('HOSTNAME', '')}",
+            f"-Dserver.cores={os.getenv('JVM_CORES', '')}",
+            "-Djava.net.preferIPv4Stack=true",
+            "-Dwork.directory=/firefly/workarea",
+            "-Dshared.work.directory=/firefly/shared-workarea",
+            "-Dserver_config_dir=/firefly/config",
+            "-Dstats.log.dir=/firefly/logs/statistics",
+            "-Dalerts.dir=/firefly/alerts",
+            f"-Dvisualize.fits.search.path={vis_path}",
+        ]
+    )
 
     # Remove admin protection if disabled
     if use_admin_auth == "false":
@@ -240,7 +272,9 @@ def main():
     os.environ["CATALINA_OPTS"] = catalina_opts
     os.environ["ADMIN_USER"] = admin_user
     os.environ["ADMIN_PASSWORD"] = admin_password
-    debug_mode = os.getenv("DEBUG", "false").lower() == "true" or (len(sys.argv) > 1 and sys.argv[1] == "--debug")
+    debug_mode = os.getenv("DEBUG", "false").lower() == "true" or (
+        len(sys.argv) > 1 and sys.argv[1] == "--debug"
+    )
     cmd = [f"{catalina_home}/bin/catalina.sh"]
     if debug_mode:
         cmd.append("jpda")
@@ -255,7 +289,7 @@ def main():
     catalina: Path = Path(catalina_home)
     webapps_ref: Path = catalina / "webapps-ref"
     webapps: Path = catalina / "webapps"
-    extract(source = webapps_ref, destination=webapps, prefix = path_prefix, kind="war")
+    extract(source=webapps_ref, destination=webapps, prefix=path_prefix, kind="war")
 
     # check for no-ops flags
     if len(sys.argv) > 1:
@@ -266,7 +300,9 @@ def main():
             dry_run(cmd, webapps)
 
     # Start background cleanup
-    subprocess.Popen([f"{catalina_home}/cleanup.sh", "/firefly/workarea", "/firefly/shared-workarea"])
+    subprocess.Popen(
+        [f"{catalina_home}/cleanup.sh", "/firefly/workarea", "/firefly/shared-workarea"]
+    )
 
     # Start Tomcat; Replace the current process with Tomcat
     print("Starting Tomcat...")
